+++ SKP_Silk_biquad.go
@@ -14,7 +14,7 @@ func SKP_Silk_biquad(in []int16, B [3]int16, A [2]int16, S []int32, out []int16)
 		S1 = SKP_SMULWB(out32, A1_neg) << 3
 		S1 = SKP_SMLABB(S1, in16, int32(B[2]))
 		tmp32 := int32(int64(SKP_RSHIFT_ROUND(out32, 13)) + 1)
-		out[k] = SKP_SAT16(int16(tmp32))
+		out[k] = SKP_SAT16(tmp32)
 	}
 	S[0] = S0
 	S[1] = S1
@@ -35,6 +35,6 @@ func SKP_Silk_biquad_alt(in []int16, B_Q28 [3]int32, A_Q28 [2]int32, S []int32, 
 		S[1] = SKP_RSHIFT_ROUND(SKP_SMULWB(out32_Q14, A1_L_Q28), 14)
 		S[1] = SKP_SMLAWB(S[1], out32_Q14, A1_U_Q28)
 		S[1] = SKP_SMLAWB(S[1], B_Q28[2], inval)
-		out[k] = SKP_SAT16(int16((int64(out32_Q14) + (1 << 14) - 1) >> 14))
+		out[k] = SKP_SAT16(int32(int64(out32_Q14)+(1<<14)-1) >> 14)
 	}
 }
